+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Snapshot tests for `Greeting`.
 - Snapshot tests for `ProductNames`.
 - Better use of helpers in tests.
+- Snapshot tests for `ProductDescription`.
 
 ## [3.17.2] - 2019-02-18